backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
nba_api==1.4.1
ffmpeg-python==0.2.0
requests==2.32.3
pytest==8.3.0
httpx==0.27.0
